Render quiz participant names as text
Quiz participant names arrive over Action Cable and are untrusted. Render them with textContent instead of innerHTML so markup is displayed literally rather than parsed and executed.

xss_foliate already sanitizes display_name before validation, but the client should not rely on that protection alone. Text-only rendering provides defense in depth if the model-level sanitization is bypassed, removed, or changed.

Extract the participant-list update into a testable module and add regression coverage for a hostile display name.

Authored by Codex (`gpt-5.6-sol xhigh`).
